Added fixes of conversion data types up on yajl definition Added Number type with atol conversion Added Casting for callbacks

+++ libgrive/src/json/ValBuilder.cc
@@ -32,7 +32,7 @@ ValBuilder::~ValBuilder()
 {
 }
 
-void ValBuilder::Visit( long long t )
+void ValBuilder::Visit( long int t )
 {
 	Build(Val(t)) ;
 }
